Skills. Explosions.
General: Coin purchase currently adds skillpoints (for development purposes).

New: Skill upgrades now save on server side.
New: Skill icons in the match are correctly set for the local player.

Fix: JS Deserialization failed to convert string to int. Conversion changed to string to float.

+++ Assets/Scripts/Services/GameService.cs
@@ -97,7 +97,7 @@ namespace Com.Hypester.DM3
             public int level;
             public int maxlevel;
             public object description;
-            public List<int> currLevel;
+            public List<float> currLevel;
         }
 
         [Serializable]
